feat: 입력요구사항 1000원 단위 예외처리
