+++ src/git.txt
@@ -4,7 +4,7 @@ git remote add origin https://github.com/jngbnss/preCoursePractice.git
 git status
 
 git add .
-git commit -m "feat: 입력값 받기"
+git commit -m "feat: 입력요구사항 1000원 단위 예외처리"
 git push origin jngbnss
 
 git pull origin main
