+++ src/main/java/lotto/Application.java
@@ -1,7 +1,9 @@
 package lotto;
 
+
 public class Application {
     public static void main(String[] args) {
-        // TODO: 프로그램 구현
+        Controller controller = new Controller();
+        controller.run();
     }
 }
