+++ src/main/java/lotto/View.java
@@ -4,4 +4,8 @@ public class View {
     public void showPrice(){
         System.out.println("구입금액을 입력해주세요.");
     }
+    public void showTickets(int ticketCount){
+
+        System.out.println("\n"+ticketCount+"개를 구매했습니다.");
+    }
 }
